simplify vite configuration down into a single file
